Update non-english.md
Changed from fake website to official.
Also, it's written AniLibria, not Anilibira. (The leader of AniLibria has often said in streams that correct writing is AniLibria)
About which is fake, which official, the leader also mentioned that any other website other than anilibria.tv and anilibria.top is fake. There's also a mirror, but it's definitely not anilibria.best