Update file-tools.md
Added image to pdf tool. I really like this because I use it often to convert comics to pdf. It works really fast, I just did a 250 png conversion to pdf and it took like 20 seconds. You don't need to download any tools or upload it to any server, it works completely offline.