Update gaming-tools.md
The wemod patcher on the github page in the undecoded link in this link is completely virus. Wemod patcher looks like open source but there is no source code of the application on the github page. My friend downloaded this application and his accounts were stolen and his mail was hijacked