Update mobile.md
Added and Starred Discoverium, Discoverium is a fork of Obtainium. Obtainium is great, but it is missing one critical feature. That feature is a search function for discovery of new apps, Discoverium then works the same as Obtainium but with the added benefit of having a search function making it work like a app store but you can get app updates straight from the source.